+++ CHANGELOG.md
@@ -9,11 +9,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Packaging
 
-- Updated shell completions
+- Minimum Rust version has been bumped to 1.45.0
 
 ### Added
 
 - IME composition preview not appearing on Windows
+- Synchronized terminal updates using `DCS = 1 s ST`/`DCS = 2 s ST`
+- Regex terminal hints ([see features.md](./docs/features.md#hints))
+
+### Fixed
+
+- Alacritty failing to start on X11 with invalid DPI reported by XRandr
+- Text selected after search without any match
+- Incorrect vi cursor position after leaving search
+
+### Removed
+
+- Config field `visual_bell`, you should use `bell` instead
+
+## 0.7.2
+
+### Packaging
+
+- Updated shell completions
 
 ### Fixed
 
@@ -21,11 +39,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Segmentation fault on shutdown with Wayland
 - Incorrect estimated DPR with Wayland
 - Consecutive clipboard stores dropped on Wayland until the application is refocused
-- Alacritty failing to start on X11 with invalid DPI reported by XRandr
-
-### Removed
-
-- Config field `visual_bell`, you should use `bell` instead
 
 ## 0.7.1